remaked DB and modified controllers
