+++ app/Http/Controllers/LabController.php
@@ -28,201 +28,6 @@ class LabController extends Controller
         $this->middleware('auth');
     }
 
-    //研究室一覧表示
-    public function index()
-    {
-        //県の画像を取得
-        $prefecture_images = Prefecture_image::get();
-
-        //メインの県名を定義し、その画像名を取得
-        $main_prefectures = [
-            ["name" => "東京都"],
-            ["name" => "大阪府"],
-            ["name" => "愛知県"],
-            ["name" => "神奈川県"],
-            ["name" => "京都府"],
-            ["name" => "福岡県"],
-        ];
-        for ($i = 0; $i < count($main_prefectures); $i++) {
-            $main_prefecture = $main_prefectures[$i];
-            $index = array_search($main_prefecture["name"], array_column($prefecture_images->toArray(), "prefecture_name"));
-            $main_prefectures[$i]["image"] = $prefecture_images[$index]->prefecture_filename;
-        }
-
-        //全県名を定義
-        $all_prefectures = [
-            [
-                "category" => "北海道・東北",
-                "prefectures" => ["北海道", "青森県", "秋田県", "山形県", "岩手県", "宮城県", "福島県"]
-            ],
-            [
-                "category" => "関東",
-                "prefectures" => ["東京都", "神奈川県", "埼玉県", "千葉県", "栃木県", "茨城県", "群馬県"]
-            ],
-            [
-                "category" => "中部",
-                "prefectures" => ["愛知県", "岐阜県", "静岡県", "三重県", "新潟県", "山梨県", "長野県", "石川県", "富山県", "福井県"]
-            ],
-            [
-                "category" => "関西",
-                "prefectures" => ["大阪府", "兵庫県", "京都府", "滋賀県", "奈良県", "和歌山県"]
-            ],
-            [
-                "category" => "中国・四国",
-                "prefectures" => ["岡山県", "広島県", "鳥取県", "島根県", "山口県", "香川県", "徳島県", "愛媛県", "高知県"]
-            ],
-            [
-                "category" => "九州・沖縄",
-                "prefectures" => ["福岡県", "佐賀県", "長崎県", "熊本県", "大分県", "宮崎県", "鹿児島県", "沖縄県"]
-            ],
-        ];
-
-        //学部のロゴを取得
-        $faculty_logos = Faculty_logo::get();
-
-        //学部名を定義し、そのロゴ画像を取得
-        $faculties = [
-            [
-                "category" => "文系学部",
-                "faculty_names" => [
-                    ["name" => "法学部"],
-                    ["name" => "経済学部"],
-                    ["name" => "文学部"],
-                    ["name" => "教育学部"],
-                    ["name" => "外国語学部"],
-                    ["name" => "その他"]
-                ]
-            ],
-            [
-                "category" => "理系学部",
-                "faculty_names" => [
-                    ["name" => "理学部"],
-                    ["name" => "工学部"],
-                    ["name" => "農学部"],
-                    ["name" => "医学部"],
-                    ["name" => "薬学部"],
-                    ["name" => "その他"]
-                ]
-            ],
-        ];
-        for ($i = 0; $i < count($faculties); $i++) {
-            $faculty_names = $faculties[$i]["faculty_names"];
-            for ($j = 0; $j < count($faculty_names); $j++) {
-                $faculty = $faculty_names[$j];
-                $index = array_search($faculty["name"], array_column($faculty_logos->toArray(), "faculty_name"));
-                $faculties[$i]["faculty_names"][$j]["image"] = $faculty_logos[$index]->faculty_filename;
-            }
-        }
-
-        //研究室の評価を取得
-        $lab_evaluations = lab_evaluation::orderBy("updated_at", "desc")->get();
-
-        //研究室の情報を取得
-        $laboratories = Laboratory::get();
-
-        //大学の情報を取得
-        $universities = Univ_data::get();
-
-        //研究室の新着口コミのデータを作成
-        $new_evaluations = [];
-        foreach (array_slice($lab_evaluations->toArray(), 0, 5) as $lab_evaluation) {
-            $id = $lab_evaluation['id'];
-            $laboratory_name = $lab_evaluation['lab_name'];
-            $university_name = $lab_evaluation['lab_univ'];
-
-            //研究室の学部を取得
-            $faculty_name = "";
-            foreach ($laboratories as $laboratory) {
-                if ($laboratory->lab_name === $laboratory_name && $laboratory->lab_univ === $university_name) {
-                    $faculty_name = $laboratory->lab_faculty;
-                }
-            }
-
-            //研究室の県を取得
-            $index = array_search($university_name, array_column($universities->toArray(), "univ_name"));
-            $prefecture_name = $universities[$index]->prefecture_name;
-
-            //県の画像を取得
-            $index = array_search($prefecture_name, array_column($prefecture_images->toArray(), "prefecture_name"));
-            $prefecture_image = $prefecture_images[$index]->prefecture_filename;
-
-            //研究室の評価を取得
-            $evaluation_items = [
-                "professor" => $lab_evaluation['prof_average'],
-                "job" => $lab_evaluation['job_average'],
-                "atmosphere" => $lab_evaluation['lab_average'],
-                "other" => $lab_evaluation['other_average']
-            ];
-
-            $new_evaluations[] = [
-                "id" => $id,
-                "prefecture_image" => $prefecture_image,
-                "laboratory_name" => $lab_evaluation['lab_name'],
-                "university_name" => $lab_evaluation['lab_univ'],
-                "faculty_name" => $faculty_name,
-                "evaluation_items" => $evaluation_items
-            ];
-        }
-
-        //高評価(all_averageが高い)の研究室を取得
-        $sort = [];
-        foreach ($lab_evaluations as $key => $value) {
-            $sort[$key] = $value["all_average"];
-        }
-        $high_evaluate_laboratories = $lab_evaluations->toArray();
-        array_multisort($sort, SORT_DESC, $high_evaluate_laboratories);
-
-        //高評価の研究室の口コミデータを作成
-        $high_evaluations = [];
-        foreach (array_slice($high_evaluate_laboratories, 0, 5) as $lab_evaluation) {
-            $id = $lab_evaluation['id'];
-            $laboratory_name = $lab_evaluation['lab_name'];
-            $university_name = $lab_evaluation['lab_univ'];
-
-            //研究室の学部を取得
-            $faculty_name = "";
-            foreach ($laboratories as $laboratory) {
-                if ($laboratory->lab_name === $laboratory_name && $laboratory->lab_univ === $university_name) {
-                    $faculty_name = $laboratory->lab_faculty;
-                }
-            }
-
-            //研究室の県を取得
-            $index = array_search($university_name, array_column($universities->toArray(), "univ_name"));
-            $prefecture_name = $universities[$index]->prefecture_name;
-
-            //県の画像を取得
-            $index = array_search($prefecture_name, array_column($prefecture_images->toArray(), "prefecture_name"));
-            $prefecture_image = $prefecture_images[$index]->prefecture_filename;
-
-            //研究室の評価を取得
-            $evaluation_items = [
-                "professor" => $lab_evaluation['prof_average'],
-                "job" => $lab_evaluation['job_average'],
-                "atmosphere" => $lab_evaluation['lab_average'],
-                "other" => $lab_evaluation['other_average']
-            ];
-
-            $high_evaluations[] = [
-                "id" => $id,
-                "prefecture_name" => $prefecture_name,
-                "prefecture_image" => $prefecture_image,
-                "laboratory_name" => $lab_evaluation['lab_name'],
-                "university_name" => $lab_evaluation['lab_univ'],
-                "faculty_name" => $faculty_name,
-                "evaluation_items" => $evaluation_items
-            ];
-        }
-
-        return view('laboratories', [
-            'main_prefectures' => $main_prefectures,
-            'all_prefectures' => $all_prefectures,
-            'faculties' => $faculties,
-            'new_evaluations' => $new_evaluations,
-            'high_evaluations' => $high_evaluations
-        ]);
-    }
-
     public function search(Request $request)
     {
         #キーワード受け取り
@@ -296,151 +101,6 @@ class LabController extends Controller
         }
     }
 
-    public function area_search($prefecture_name)
-    {
-        $categories = [
-            [
-                'name' => '総合評価',
-                'key' => 'all_average'
-            ],
-            [
-                'name' => '教授',
-                'key' => 'prof_average'
-            ],
-            [
-                'name' => '就活',
-                'key' => 'job_average'
-            ],
-            [
-                'name' => '雰囲気',
-                'key' => 'lab_average'
-            ],
-            [
-                'name' => 'その他',
-                'key' => 'other_average'
-            ],
-        ];
-
-        //検索された文字列をprefecture_nameカラムに含む大学名の一覧を取得
-        $university_names = Univ_data::where('prefecture_name', 'like', "%$prefecture_name%")->get('univ_name');
-
-        //検索条件に当てはまる大学の研究室評価の平均値を取得
-        $average_evaluations = lab_evaluation::selectRaw('lab_univ, AVG(all_average) as all_average, AVG(prof_average) as prof_average, AVG(job_average) as job_average, AVG(lab_average) as lab_average, AVG(other_average) as other_average')->whereIn('lab_univ', $university_names)->groupBy('lab_univ')->get()->toArray();
-
-        //大学ごとに評価を取得
-        $universities = array_map(
-            function ($average_evaluation) use ($categories) {
-                $university_name = $average_evaluation['lab_univ'];
-
-                //note: ここで改めてクエリを発行しているが、$average_evaluationsを取得するときに同時に取得した方がデータベースへのアクセスが1回で済むのでいいかも。
-                //新着口コミを取得
-                $latest_evaluation = lab_evaluation::latest()->where('lab_univ', $university_name)->first()->toArray();
-                $laboratory_name = $latest_evaluation['lab_name'];
-
-                return [
-                    'name' => $university_name,
-                    'averageTotalEvaluation' => [
-                        'category' => $categories[0]['name'],
-                        'value' => round($average_evaluation[$categories[0]['key']], 2)
-                    ],
-                    'averageEvaluations' => array_map(function ($category) use ($average_evaluation) {
-                        return [
-                            'category' => $category['name'],
-                            'value' => round($average_evaluation[$category['key']], 2)
-                        ];
-                    }, array_slice($categories, 1)),
-                    'latestEvaluation' => [
-                        'laboratoryName' => $laboratory_name,
-                        'facultyName' => Laboratory::where('lab_univ', $university_name)->where('lab_name', $laboratory_name)->value('lab_faculty'),
-                        'evaluationValues' =>
-                            array_map(function ($category) use ($latest_evaluation) {
-                                return [
-                                    'category' => $category['name'],
-                                    'value' => round($latest_evaluation[$category['key']], 2)
-                                ];
-                            }, $categories),
-                    ]
-                ];
-            }, $average_evaluations);
-
-        return view('area_search_result', [
-            'prefectureName' => $prefecture_name,
-            'universities' => $universities
-        ]);
-    }
-
-    /*======== 学部ごとに探すページ =================*/
-    public function faculty_result($faculty)
-    {
-        $keyword = $faculty;
-
-        $faculty_logos = faculty_logo::orderBy('created_at', 'asc')->get();
-        $average_item_jp = ["総合評価", "教授", "就活", "研究室", "その他"];
-
-        $data = array();
-
-        if ($keyword == "その他") {
-            $faculty_array = [
-                ["法学部", "経済学部", "文学部", "教育学部", "外国語学部"],
-                ["理学部", "工学部", "農学部", "医学部", "薬学部",],
-            ];
-            $laboratories = Laboratory::latest()->whereNotIn('lab_faculty', $faculty_array[0])->whereNotIn('lab_faculty', $faculty_array[1])->get();
-            $isLab = Laboratory::latest()->whereNotIn('lab_faculty', $faculty_array[0])->whereNotIn('lab_faculty', $faculty_array[1])->count();
-
-        } else {
-            $laboratories = Laboratory::latest()->where('lab_faculty', $faculty)->get();
-            $isLab = Laboratory::where('lab_faculty', $keyword)->count();
-
-        }
-
-        if ($isLab != 0) {
-            /*=====研究室ごとの平均・最新口コミ評価計算部===========*/
-            $average_item = ['all_average', 'prof_average', 'job_average', 'lab_average', 'other_average'];
-            $count = 0;
-
-            //”○○”学部で絞った研究室情報(laboratories)から、１つずつとりだし、それに対応するlab_evaluationを取り出す
-            foreach ($laboratories as $laboratory) {
-                $counter = lab_evaluation::where('lab_univ', $laboratory->lab_univ)->where('lab_name', $laboratory->lab_name)->count();
-                if ($counter > 1) {
-                    for ($n = 0; $n < 5; $n++) {
-                        //平均口コミ
-                        $array_average[$laboratory->id][$n] = lab_evaluation::where('lab_univ', $laboratory->lab_univ)
-                            ->where('lab_name', $laboratory->lab_name)
-                            ->avg($average_item[$n]);
-                        //最新口コミ
-                        $array_latest_evaluation[$laboratory->id] = lab_evaluation::latest()
-                            ->where('lab_univ', $laboratory->lab_univ)
-                            ->where('lab_name', $laboratory->lab_name)
-                            ->first();
-                    }
-                } else if ($counter == 1) {
-                    $tmp_evaluation = lab_evaluation::where('lab_univ', $laboratory->lab_univ)->where('lab_name', $laboratory->lab_name)->first();
-                    for ($n = 0; $n < 5; $n++) {
-                        $tmp = $average_item[$n];
-                        //平均口コミ
-                        $array_average[$laboratory->id][$n] = $tmp_evaluation->$tmp;
-                        //最新口コミ
-                        $array_latest_evaluation[$laboratory->id] = $tmp_evaluation;
-                    }
-                }
-            }
-            return view('faculty_result', [
-                'keyword' => $keyword,
-                'laboratories' => $laboratories,
-                'faculty_logos' => $faculty_logos,
-                'average_item_jp' => $average_item_jp,
-                'faculty' => $faculty,
-                'array_average' => $array_average,
-                'array_latest_evaluation' => $array_latest_evaluation,
-            ]);
-        } else   //該当データなしの場合
-        {
-            return view('faculty_result', [
-                'keyword' => $keyword,
-                'laboratories' => $laboratories,
-            ]);
-        }
-    }
 
 
     //研究室情報を追加する
