+++ database/factories/UserFactory.php
@@ -18,11 +18,23 @@ use Illuminate\Support\Str;
 */
 
 $factory->define(User::class, function (Faker $faker) {
+    for ($i = 1; $i < 21; $i++)
+    {
+      $random_date = [rand(2017, 2019), rand(1, 12), rand(1,31)];
+
+      if(!checkdate($random_date[1], $random_date[2], $random_date[0])){
+        $random_date = [rand(2017, 2019), rand(1, 12), 1];
+      }
+    }
+
     return [
-        'name' => $faker->name,
         'email' => $faker->unique()->safeEmail,
-        'email_verified_at' => now(),
         'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
-        'remember_token' => Str::random(10),
+        'role' => 10,
+        'created_at'         => new DateTime($random_date[0].'-'.$random_date[1].'-'.$random_date[2]),
+        'updated_at'         => new DateTime($random_date[0].'-'.$random_date[1].'-'.$random_date[2]),
+        'token'              => str_random(15),
+        'email_verified_at' => now(),
+        'status'             => 1,
     ];
 });
