+++ resources/views/components/botton_watch_reviews.blade.php
@@ -1,5 +1,5 @@
 <td>
-  <a class="nav-link" href="{{ url('lab/'.$laboratory->lab_univ.'/'.$laboratory->lab_name) }}">
+  <a class="nav-link" href="{{ url('lab/'.$laboratory[0]->lab_univ.'/'.$laboratory[0]->lab_name) }}">
     この研究室の口コミを見る
   </a>
 </td>
