+++ resources/views/components/faculty_logo.blade.php
@@ -1,14 +1,6 @@
 <td class="table-text">
 　  <!-- 学部名 -->
-  {{ $laboratory->lab_faculty }}
+  {{$laboratory[1]->faculty_name}}
   <!-- 学部ロゴ表示 -->
-  @foreach($faculty_logos as $faculty_logo)
-    @if($faculty_logo->faculty_name == $laboratory->lab_faculty)
-      <img src="{{ asset('img/fac_logo/' .$faculty_logo->faculty_logo) }}" alt="{{ $faculty_logo->faculty_logo }}" width="75" height="75">
-    @break
-    @else
-      <img src="{{ asset('img/fac_logo/other.png') }}" alt="{{ $faculty_logo->faculty_logo }}" width="75" height="75">
-    @break
-    @endif
-  @endforeach
+  <img src="{{ asset('img/faculty_logo/' .$laboratory[1]->faculty_filename) }}" alt="{{ $laboratory[1]->faculty_filename}}" width="75" height="75">
 </td>
