+++ resources/views/top.blade.php
@@ -2,4 +2,3 @@
 @section('content')
     <top :areas='@json($areas)' :route-register='@json(route('register'))'></top>
 @endsection
-
